Moved Supporters components to own folder
